# src/OpenApiMcp.Server/demo/demoapi.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "openapi-mcp bundled demo API (synthetic, offline)",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:5088" }],
  "paths": {
    "/posts": {
      "get": {
        "operationId": "listPosts",
        "summary": "List posts, optionally filtered by userId and paginated.",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "description": "Filter posts by author id.",
            "schema": { "type": "integer" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number (1-based). Responses include a Link rel=\"next\" header when more pages exist.",
            "schema": { "type": "integer" }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "description": "Items per page (1-100).",
            "schema": { "type": "integer" }
          }
        ]
      },
      "post": {
        "operationId": "createPost",
        "summary": "Create a post (MUTATING — disabled unless the policy allows it).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "body": { "type": "string" },
                  "userId": { "type": "integer" }
                }
              }
            }
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "operationId": "getPost",
        "summary": "Fetch a single post by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The post id.",
            "schema": { "type": "integer" }
          }
        ]
      }
    }
  }
}